Lots of cleanup
- Removed unused svg icons from an old version of the playground page
- Changed from using code-prettify to just using codemirror's `runMode`
function for formatting code samples on the homepage/docs page. This
makes all the code styles consistent, and allowed me to remove a few
dependencies and clean up some code. It also means that new syntax
features don't need to be implemented twice, once for the playground
page and once for the docs pages
- Begin scaffolding for Manual (new docs) page