Handful of changes
- Some minor styling tweaks in mobile (lesser padding)
- Lazily load the routes to keep package size small
- Prepare for markdown-powered documentation pages in the secret
`/manual` route. This also is the reason for the `config-overrides.js`
file being pulled in; CRA by default wants to load `.md` files using
`file-loader`, but that would require a client-side http request to load
in the markdown file. I want the markdown file to be baked into the page
so the resulting react component can be built synchronously.